Implement loopback network.
